requirements.txt
Django==2.0.7
django-crispy-forms==1.7.2
django-markdown-deux==1.0.5
django-pagedown==1.0.5
markdown2==2.3.5
Pillow==5.2.0
pytz==2018.5
